# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: |
    Tree Friend E-Commerce API.

    ## Versioning

    The API is mounted under two prefixes:
      - `/api/v1` — the canonical, versioned path. New clients should use this.
      - `/api`     — backward-compat alias for existing callers. Same handlers,
                     transparently aliased.

    When a v2 is needed, a new router will be mounted under `/api/v2` and
    `/api/v1` will be left untouched. The unversioned `/api` alias will be
    deprecated (Deprecation header + warning log) once all callers have
    migrated to `/api/v1`.
servers:
  - url: /api/v1
    description: Versioned API path (canonical)
  - url: /api
    description: Unversioned API path (backward-compat alias)
...
paths:
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List current user's orders
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place an order
      description: >-
        Always returns an array. A cart spanning multiple sellers splits
        into one order per seller (plus one for admin-direct lines, if
        any) -- plan doc §2, §7. A single-seller or all-admin-direct cart
        still returns an array, just with one element, so callers never
        need to branch on whether checkout split.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Created order(s)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
...
components:
  schemas:
...
    OrderLineItem:
      type: object
      description: >-
        Snapshot of a purchased line item, stored as JSON on the order at
        checkout time. Either variantId+variantName are set (admin-direct
        line) or sellerListingId+sellerId are set (marketplace line) --
        never both. Every line on a given order has the same kind and, for
        marketplace orders, the same sellerId as the order itself.
      properties:
        productId:
          type: integer
        productName:
          type: string
        productImage:
          type: string
        variantId:
          type: integer
        variantName:
          type: string
        sellerListingId:
          type: integer
        sellerId:
          type: integer
        quantity:
          type: integer
        price:
          type: number
        deliveryCharge:
          type: number
      required:
        - productId
        - productName
        - productImage
        - quantity
        - price
        - deliveryCharge
    Order:
      type: object
      properties:
        id:
          type: integer
        trackingId:
          type: string
        userId:
          type: string
        sellerId:
          type: ["integer", "null"]
          description: Null for admin-direct orders; set for marketplace orders.
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderLineItem"
        totalAmount:
          type: number
        paymentMethod:
          type: string
        paymentStatus:
          type: string
        senderNumber:
          type: ["string", "null"]
        paidAt:
          type: ["string", "null"]
        orderStatus:
          type: string
        transactionId:
          type: ["string", "null"]
        shippingAddress:
          $ref: "#/components/schemas/AddressBody"
        couponCode:
          type: ["string", "null"]
        discountAmount:
          type: number
        cancellationReason:
          type: ["string", "null"]
        giftWrap:
          type: string
        giftMessage:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - trackingId
        - userId
        - items
        - totalAmount
        - paymentMethod
        - paymentStatus
        - orderStatus
        - transactionId
        - shippingAddress
        - couponCode
        - discountAmount
        - createdAt
        - updatedAt
...
    CreateOrderBody:
      type: object
      description: >-
        A cart spanning multiple sellers splits into one order per seller
        group (plan doc §2, §7). paymentMethod is the fallback/default used
        for any group not present in sellerPaymentMethods; a single-seller
        or all-admin-direct cart can just send paymentMethod and omit
        sellerPaymentMethods entirely.
      properties:
        paymentMethod:
          type: string
          description: Fallback payment method for any seller group not present in sellerPaymentMethods.
        sellerPaymentMethods:
          type: object
          additionalProperties:
            type: string
          description: >-
            Per-seller-group payment method override, keyed by sellerId as
            a string ("null" for the admin-direct group). Each seller only
            accepts the payment methods enabled on their listings (plan doc
            §7) -- the buyer picks per group at checkout, not once
            globally.
        transactionId:
          type: ["string", "null"]
        senderNumber:
          type: ["string", "null"]
          description: >-
            Fallback bKash sending number for any seller group not present
            in sellerSenderNumbers. Required when the resolved payment
            method for a group is "bkash" and that group has no override
            in sellerSenderNumbers.
        sellerSenderNumbers:
          type: object
          additionalProperties:
            type: ["string", "null"]
          description: >-
            Per-seller-group bKash sending number override, keyed by
            sellerId as a string ("null" for the admin-direct group) --
            same key convention as sellerPaymentMethods. Added in Part 5:
            previously a single top-level senderNumber was reused across
            every seller group resolving to "bkash", which doesn't hold up
            once a cart has multiple sellers with separate bKash merchant
            accounts (PHASE3_HANDOFF.md flagged this as a known gap).
        shippingAddress:
          $ref: "#/components/schemas/AddressBody"
        couponCode:
          type: ["string", "null"]
        loyaltyPointsToRedeem:
          type: integer
        giftWrap:
          type: boolean
        giftMessage:
          type: ["string", "null"]
      required:
        - shippingAddress
...
    AddressBody:
      type: object
      properties:
        fullName:
          type: string
        phone:
          type: string
        street:
          type: string
        city:
          type: string
        district:
          type: string
        postalCode:
          type: ["string", "null"]
        isDefault:
          type: boolean
      required:
        - fullName
        - phone
        - street
        - city
        - district
...